fix(admin): redirect authed admins from /admin/login server-side
The client-side useEffect redirect was unreliable — after a successful
sign-in the user landed on /admin/login with a session, sidebar rendered
(layout took the authed branch), but the page itself never bounced to
/admin so the dashboard never appeared.

Split the route into a server page that calls `auth()` and `redirect("/admin")`
when the session is admin-tier, plus a client form component for the
credentials submit and the "signed in but not an admin" recovery state.
Server redirect runs before any markup ships, so the dashboard appears
on first paint instead of after a hydration race.
